tests/api_testing/requirements.txt
requests>=2.31.0
pytest>=7.4.0
hypothesis>=6.92.0
pyyaml>=6.0.1
pydantic>=2.6.0
colorama>=0.4.6
